// src/app/model/filter-model.ts
export class Gender {
  name: string = '';
  keywords: string = '';
}

export class FilterModel {
  genders: Gender[] = [];
  category: CategoryNode[] = [];
  store: Store[] = [];
  min_price: string = '0';
  max_price: string = '10000';
  brand: Brand[] = [];
}

export class CategoryNode {
  id: string = '';
  data: any;
  count: number = 0;
  name: string = '';
  children: CategoryNode[] = [];
}